+++ .changeset/many-foxes-taste.md
@@ -3,3 +3,6 @@
 ---
 
 Add `DELETE_API_KEY_PAIR_ERROR` to `TurnkeyErrorCodes`
+Add `INITIALIZE_API_KEY_STAMPER_ERROR` to `TurnkeyErrorCodes`
+Add `INITIALIZE_PASSKEY_STAMPER_ERROR` to `TurnkeyErrorCodes`
+Add `INITIALIZE_WALLET_MANAGER_ERROR` to `TurnkeyErrorCodes`
