+++ packages/core/src/utils.ts
@@ -419,6 +419,20 @@ export const toExternalTimestamp = (
   };
 };
 
+export const ERC20_TRANSFER_ABI = [
+  {
+    type: "function",
+    name: "transfer",
+    stateMutability: "nonpayable",
+    inputs: [
+      { name: "to", type: "address" },
+      { name: "amount", type: "uint256" },
+    ],
+    outputs: [{ name: "success", type: "bool" }],
+  },
+] as const;
+
+
 export async function getActiveSessionOrThrowIfRequired(
   stampWith: StamperType | undefined,
   getActiveSession: () => Promise<Session | undefined>,
