+++ packages/react-wallet-kit/src/providers/client/Provider.tsx
@@ -24,7 +24,9 @@ import {
   storeOAuthAddProviderMetadata,
   storePKCEVerifier,
 } from "../../utils/oauth";
+import { base64UrlToBase64, atob } from "@turnkey/encoding";
 import {
+  AUTHENTICATOR_TRANSPORT_MAP,
   isValidSession,
   mergeWalletsWithoutDuplicates,
   SESSION_WARNING_THRESHOLD_MS,
@@ -1286,7 +1288,26 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       setSession(session);
       setAllSessions(allSessions);
 
-      await Promise.all([maybeRefreshWallets(), maybeRefreshUser()]);
+      const [, user] = await Promise.all([maybeRefreshWallets(), maybeRefreshUser()]);
+
+      if (user && client?.config?.passkeyConfig) {
+        const allowCredentials: PublicKeyCredentialDescriptor[] =
+          user.authenticators.map((a) => {
+            const b64 = base64UrlToBase64(a.credentialId);
+            const binary = atob(b64);
+            const id = Uint8Array.from(binary, (c: string) => c.charCodeAt(0));
+            return {
+              id,
+              type: "public-key",
+              transports: a.transports
+                .map((t) => AUTHENTICATOR_TRANSPORT_MAP[t])
+                .filter(Boolean) as AuthenticatorTransport[],
+            };
+          });
+        await client.overridePasskeyStamper({
+          config: { ...client.config.passkeyConfig, allowCredentials },
+        });
+      }
 
       if (
         masterConfig?.auth?.verifyWalletOnSignup === true &&
@@ -1560,7 +1581,7 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
   );
 
   const refreshUser = useCallback(
-    async (params?: RefreshUserParams): Promise<void> => {
+    async (params?: RefreshUserParams): Promise<v1User | undefined> => {
       const { stampWith, organizationId, userId } = params || {};
 
       if (!client)
@@ -1584,6 +1605,8 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
       if (user) {
         setUser(user);
       }
+
+      return user;
     },
     [client, callbacks, fetchUser, logout],
   );
@@ -1598,8 +1621,8 @@ export const ClientProvider: React.FC<ClientProviderProps> = ({
    * @returns A promise that resolves when the user is refreshed, or does nothing if auto-refresh is disabled
    */
   const maybeRefreshUser = useCallback(
-    async (params?: RefreshUserParams): Promise<void> => {
-      if (!masterConfig?.autoRefreshManagedState) return;
+    async (params?: RefreshUserParams): Promise<v1User | undefined> => {
+      if (!masterConfig?.autoRefreshManagedState) return undefined;
       return refreshUser(params);
     },
     [masterConfig, refreshUser],
